Integer literals reach INT64_MAX (#106)
You could not write an i64 literal. The lexer capped every base at INT32_MAX
and the AST carried the value in an int32_t, so the only way to get a number
past 2^31 was to compute one.

The ceiling is now INT64_MAX, in all three bases, with the token payload and
`lit_int.value` widened to match. One past INT64_MAX is still a lex error
rather than a promotion: the tower goes to bigint next and that tier does not
exist yet (slice 2). `-9223372036854775808` stays out of reach for the same
reason C needs `LLONG_MIN` — the digits are lexed before the sign folds in.

A literal that fits i32 lowers exactly as before, to the inline i32 constant.
A wider one has no inline form (a dynamic wide int lives on the heap), so
codegen builds it through `jacl_i64_box` — the same canonicalizing box a typed
i64 uses crossing to dyn (#107). That is what lets a wide literal be a map key:
it hashes like every other spelling of the same number. Every constant-JaclVal
site goes through one helper now, so indices and dot keys widen with it.

The typer types an out-of-i32 literal i64. Deliberately also under an i32
*expectation*, because an i32 expectation is not always a user annotation — a
binop unifies its operands by expecting the first one's type, which is what
`[- 0 9223372036854775807]` is.

Two places still carry the literal as an i32 and now refuse a wider one
instead of narrowing it silently: the staged-syntax form (#113) and a
`[Buf N T]` length, which no buffer could reach anyway.

Verified: build.sh 25/25, harness codegen 109/109, corpus 434/436 browser-pass
(the two are the pre-existing tour.jacl emit-fail and io_pipe_to_file
wrong-output; 433/435 before this change, plus the new case).

+++ src/gc.c
@@ -1433,7 +1433,7 @@ typedef struct {
     uint32_t  scope_mark;  /* hygiene: 0 = no macro context */
     union {
         struct { JaclVal head; JaclVal args; }  command;       /* head: syntax, args: vec of syntax */
-        struct { int32_t value; }               lit_int;
+        struct { int64_t value; }               lit_int;
         struct { float   value; }               lit_float;
         struct { JaclVal value; }               lit_string;    /* JaclVal string */
         struct { JaclVal name; }                var_ref;       /* JaclVal string */
